+++ web/project1/updateMeeting.php
@@ -78,9 +78,11 @@ ini_set('display_errors', 1);
 	  
 	  $studentsUpdate = explode("\n", $text); 
 	  
-	  $studentId = 0;
+	  
 	  foreach($studentsUpdate as $student)
 	  {
+		  $studentId = 0;
+		  
 		  if ($student != null)
 		  {
 			$name = explode(" ", $student);
@@ -95,7 +97,11 @@ ini_set('display_errors', 1);
 		  
 			$row = $statement2->fetch(PDO::FETCH_ASSOC);
 			$studentId = $row['id'];	
-          	  
+          	 
+			if ($studentId == 0)
+			{				  
+		      include_once('studentNotFound.php');     
+			}
 	  
 			$statement3 = $db->prepare("INSERT INTO meeting_attendance (meeting_id, student_id)
 	                              VALUES (:meetingId, :studentId)");
@@ -110,9 +116,10 @@ ini_set('display_errors', 1);
 	  $text2 = validate($_POST['students_remove']);
 	  $studentsRemove = explode("\n", $text2);
 	  
-	  $removeId = 0;
 	  foreach($studentsRemove as $student)
 	  {
+		  $removeId = 0;
+		  
 		  if ($student != null)
 		  {
 			$name = explode(" ", $student);
@@ -127,13 +134,16 @@ ini_set('display_errors', 1);
 			$stmnt->execute();
 		   
 			$row = $stmnt->fetch(PDO::FETCH_ASSOC);
-			$studentId = $row['id'];
-		  
-			$stmnt2 = $db->prepare('DELETE FROM meeting_attendance WHERE
+			$removeId = $row['id'];
+			
+			if ($removeId != 0)
+			{		  
+			  $stmnt2 = $db->prepare('DELETE FROM meeting_attendance WHERE
 								student_id = :studentId AND meeting_id = :meetingId');
-			$stmnt2->bindValue(':studentId', $studentId);
-			$stmnt2->bindValue(':meetingId', $meetingId);
-			$stmnt2->execute();
+			  $stmnt2->bindValue(':studentId', $studentId);
+			  $stmnt2->bindValue(':meetingId', $meetingId);
+			  $stmnt2->execute();
+			}
 		  }
 	  }
